fix: Clear cached rename schema (#15902)
