+++ crates/polars-plan/src/logical_plan/optimizer/projection_pushdown/mod.rs
@@ -649,10 +649,7 @@ impl ProjectionPushDown {
                     schema: Arc::new(new_schema),
                 })
             },
-            MapFunction {
-                input,
-                ref function,
-            } => functions::process_functions(
+            MapFunction { input, function } => functions::process_functions(
                 self,
                 input,
                 function,
